Third Installation Method Added
Added the CMDProcessor library and found in AOKP, providing a third fallback
installation method for VIPER4Android.

Now three methods will be used if needed:
  1. RootTools
  2. vbox
  3. CMDProcessor

The CMDProcessor method does not use any of the v4a specific methods for
shell operations such as OpenRootShell which I have seen cause issues in
a number of ROMS.

With more time I believe the CMDProcessor method could be used to replace
all other methods as it is has been tested with AOKP roms for quite sometime.