AudioManager & Code Style Updates
Added the frameworks_intermediates class file to allow access to the hidden
AudioManager API's, which allows access to the system method for creating
a Audio Effect class instead of using reflection.

Also allows access to the AudioManager methods to convert ints to byte[] and to
concat arrays.

Further code formatting as per Android AOSP guidelines.